requirements.txt
# Core ML/Data
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Streamlit & Visualization
streamlit>=1.28.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Explainability & Causal
shap>=0.42.0
dowhy>=0.9.0
causal-learn>=0.1.3

# Utils & Productivity
pyyaml>=6.0
joblib>=1.3.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
jupyter>=1.0.0
ipykernel>=6.0.0